Created frontend of hardware page. Additionally changed structure of entrire application.  Created Base Service and Base Controller.

+++ backend/src/main/java/de/mightypc/backend/model/PC.java
@@ -3,6 +3,7 @@ package de.mightypc.backend.model;
 import de.mightypc.backend.model.specs.PcCase;
 import org.springframework.data.annotation.Id;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record PC(
@@ -11,9 +12,9 @@ public record PC(
         String name,
         Specs specs,
         PcCase pcCase,
-        float price
+        BigDecimal price
 ) {
-    public PC(String name, Specs specs, PcCase pcCase, float price){
+    public PC(String name, Specs specs, PcCase pcCase, BigDecimal price){
         this(UUID.randomUUID().toString(), name, specs, pcCase, price);
     }
 }
